refactor(bridge): decouple from deprecated scitex.canvas.model
Rewire bridge module + bridge tests to import figure-model dataclasses
from scitex.io.bundle.kinds._plot._models (the canonical home) instead
of scitex.canvas.model (deprecated). The two were near-identical
duplicates; pointing at the bundle copy frees scitex.canvas of external
dataclass dependents so it can be retired as a self-contained unit
later.

No behavior change. Bridge tests pass (30 passed, 8 xfailed
pre-existing).
